Update flask_app.py
- qiy_nodes_events_source
Fix for not receiving server-sent events in pythonanywhere:

Instructions from pythonanywhere:

SSE works fine on PythonAnywhere with some things you need to know and some caveats:
Things you need to know:
You need to set the X-Accel-Buffering header to no on your event-stream response.

Caveats:
The generation of the event stream uses up a worker. On a Free account, you only have one worker, so if you use an infinite stream, your worker would never be available to handle any other requests, so the best way to use the event-stream is really to only send one message and then let the automatic rebuilding of the connection connect for the next message.
Also, requests that last more than 5 minutes will be interpreted by our servers as a broken web app, so the web app will be terminated and restarted, so it would be a good idea to ensure that the event stream never lasts more than 5 minutes - this shouldn't be a problem if you're using the "one message or nothing and done" mechanism I describe above.

As an example, here is a full flask app that implements SSEs. It does not fully implement the "one message or none and done" process I outlined above - It will tie up a worker for 10s at a time and then release it.

from flask import Flask, Response
import time

app = Flask(__name__)

@app.route('/')
def root():
    return """
    <html>
    <script>
    console.log("Setting up events");
    const eventSource = new EventSource("http://glenn.pythonanywhere.com/events/");
    eventSource.onmessage = (e) => {
      console.log(e);
    };
    console.log("Done setting up events");
    </script>
    </html>
    """

@app.route("/events/")
def events():
    def event_stream():
        ii = 0
        for ii in range(10):
            yield "data: {}\n\n".format(ii)
            ii += 1
            time.sleep(1)
    response = Response(event_stream(), mimetype="text/event-stream")
    response.headers['X-Accel-Buffering'] = 'no'
    return response